refactor(nav-drawer): consume colors from navdrawer schema
