+++ src/components/nav-drawer/nav-drawer-item.ts
@@ -7,9 +7,11 @@ import {
 } from 'lit/decorators.js';
 import { themes } from '../../theming/theming-decorator.js';
 import { partNameMap } from '../common/util.js';
-import { styles as fluent } from './themes/light/fluent/nav-drawer-item.fluent.css.js';
-import { styles as indigo } from './themes/light/indigo/nav-drawer-item.indigo.css.js';
-import { styles } from './themes/light/nav-drawer-item.base.css.js';
+import { styles } from './themes/light/item/nav-drawer-item.base.css.js';
+import { styles as fluent } from './themes/light/item/nav-drawer-item.fluent.css.js';
+import { styles as indigo } from './themes/light/item/nav-drawer-item.indigo.css.js';
+import { styles as bootstrap } from './themes/light/item/nav-drawer-item.bootstrap.css.js';
+import { styles as material } from './themes/light/item/nav-drawer-item.material.css.js';
 
 /**
  * Represents a navigation drawer item.
@@ -23,7 +25,7 @@ import { styles } from './themes/light/nav-drawer-item.base.css.js';
  * @csspart icon - The icon container.
  * @csspart content - The content container.
  */
-@themes({ fluent, indigo })
+@themes({ fluent, indigo, bootstrap, material })
 export default class IgcNavDrawerItemComponent extends LitElement {
   public static readonly tagName = 'igc-nav-drawer-item';
   public static override styles = styles;
