--- src/components/nav-drawer/themes/dark/nav-drawer-item.fluent.scss
@@ -1,15 +1,0 @@
-@use 'styles/utilities' as *;
-@use 'themes' as *;
-
-@mixin theme() {
-    igc-nav-drawer {
-        @include css-vars-from-theme($fluent, 'ig-navdrawer');
-    }
-
-    igc-nav-drawer-item {
-        --hover-background: #{color(gray, 100, .5)};
-        --active-background: #{color(primary, 400, .12)};
-        --active-icon-color: #{color(primary, 100)};
-        --active-text-color: #{color(primary, 200)};
-    }
-}
