--- src/components/nav-drawer/themes/dark/nav-drawer-item.indigo.scss
@@ -1,14 +1,0 @@
-@use 'styles/utilities' as *;
-@use 'themes' as *;
-
-@mixin theme() {
-    igc-nav-drawer {
-        @include css-vars-from-theme($indigo, 'ig-navdrawer');
-    }
-
-    igc-nav-drawer-item {
-        --text-color: #{color(gray, 100)};
-        --hover-text-color: #{color(gray, 50)};
-        --active-background: #{color(gray, 400, .5)};
-    }
-}
