--- src/components/nav-drawer/themes/light/fluent/nav-drawer-item.fluent.scss
@@ -1,15 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../themes' as *;
-
-$theme: $fluent;
-$active-icon-color: var(--active-icon-color, color(primary, 500)) !default;
-
-[part='content'] {
-    @include type-style('body-2');
-}
-
-:host([active]) {
-    [part='icon'] {
-        color: $active-icon-color;
-    }
-}
