+++ src/components/nav-drawer/themes/light/header-item/nav-drawer-header-item.base.scss
@@ -4,7 +4,4 @@
     display: block;
     padding: rem(20px) rem(16px);
     white-space: nowrap;
-    color: color(gray, 800);
-
-    @include type-style('subtitle-1');
 }
