+++ src/components/nav-drawer/themes/light/header-item/nav-drawer-header-item.fluent.scss
@@ -1,0 +1,10 @@
+@use 'styles/utilities' as *;
+@use '../themes' as *;
+
+$theme: $fluent;
+
+:host {
+    @include type-style('subtitle-1');
+
+    color: var-get($theme, 'item-header-text-color');
+}
