--- src/components/nav-drawer/themes/light/indigo/nav-drawer-header-item.indigo.scss
@@ -1,8 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../themes' as *;
-
-$theme: $indigo;
-
-:host {
-    color: color(gray, 50);
-}
