--- src/components/nav-drawer/themes/light/indigo/nav-drawer-item.indigo.scss
@@ -1,36 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../themes' as *;
-
-$theme: $indigo;
-$active-background: var(--active-background, color(gray, 600, .8)) !default;
-$text-color: var(--text-color, color(gray, 300, .8)) !default;
-$hover-text-color: var(--hover-text-color, color(gray, 100)) !default;
-
-:host {
-    @include css-vars-from-theme($theme, 'ig-nav-drawer');
-
-    [part~='base'] {
-        border-radius: var-get($theme, 'item-border-radius');
-        gap: rem(24px);
-        color: $text-color;
-        padding: rem(4px);
-        margin: rem(15px) rem(8px);
-
-        &:hover,
-        &:focus-within {
-            background: transparent;
-            color: $hover-text-color;
-        }
-    }
-
-    [part='content'] {
-        padding-inline-start: rem(8px);
-    }
-}
-
-:host([active]) {
-    [part~='base'] {
-        background: $active-background;
-        color: color(gray, 50);
-    }
-}
