--- src/components/nav-drawer/themes/light/indigo/nav-drawer.indigo.scss
@@ -1,21 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../themes' as *;
-
-$theme: $indigo;
-$background: var(--background, color(gray, 700)) !default;
-
-:host {
-    @include css-vars-from-theme($theme, 'ig-nav-drawer');
-
-    [part='base'] {
-        border-radius: var-get($theme, 'border-radius');
-        background: $background;
-        border: none;
-        box-shadow: none;
-    }
-}
-
-:host([position='start']) [part='mini'],
-:host([position='end']) [part='mini'] {
-    background: $background;
-}
