+++ src/components/nav-drawer/themes/light/item/nav-drawer-item.base.scss
@@ -1,23 +1,11 @@
 @use 'styles/utilities' as *;
-@use 'themes' as *;
-
-$theme: $material;
-$active-text-color: var(--active-text-color, color(primary, 500)) !default;
-$active-background: var(--active-background, color(primary, 500, .12)) !default;
-$hover-background: var(--hover-background, color(gray, 200, .8)) !default;
-
-:host {
-    @include css-vars-from-theme($theme, 'ig-nav-drawer');
-}
 
 [part~='base'] {
-    border-radius: var-get($theme, 'item-border-radius');
     display: flex;
     gap: rem(32px);
     align-items: center;
     flex-flow: row nowrap;
     position: relative;
-    color: color(gray, 700);
     max-height: rem(48px);
     margin: rem(8px);
     padding: rem(12px) rem(8px);
@@ -34,17 +22,10 @@ $hover-background: var(--hover-background, color(gray, 200, .8)) !default;
         display: flex;
         font: inherit;
     }
-
-    &:hover,
-    &:focus-within {
-        background: $hover-background;
-        color: color(gray, 900);
-    }
 }
 
 [part='content'] {
     @include ellipsis();
-    @include type-style('subtitle-2');
 }
 
 [part='base mini'] {
@@ -63,15 +44,4 @@ $hover-background: var(--hover-background, color(gray, 200, .8)) !default;
     user-select: none;
     cursor: initial;
     pointer-events: none;
-
-    [part~='base'] {
-        color: color(gray, 400);
-    }
-}
-
-:host([active]) {
-    [part~='base'] {
-        background: $active-background;
-        color: $active-text-color;
-    }
 }
