+++ src/components/nav-drawer/themes/light/item/nav-drawer-item.indigo.scss
@@ -1,0 +1,38 @@
+@use 'styles/utilities' as *;
+@use '../themes' as *;
+
+$theme: $indigo;
+$active-background: var(--active-background, color(gray, 600, .8)) !default;
+$text-color: var(--text-color, color(gray, 300, .8)) !default;
+$hover-text-color: var(--hover-text-color, color(gray, 100)) !default;
+
+:host {
+    @include css-vars-from-theme($theme, 'ig-nav-drawer');
+}
+
+[part~='base'] {
+    gap: rem(24px);
+    padding: rem(4px);
+    margin: rem(15px) rem(8px);
+    border-radius: var-get($theme, 'item-border-radius');
+    color: var-get($theme, 'item-text-color');
+
+    &:hover,
+    &:focus-within {
+        background: var-get($theme, 'item-hover-background');
+        color: var-get($theme, 'item-hover-text-color');
+    }
+}
+
+[part='content'] {
+    @include type-style('subtitle-2');
+
+    padding-inline-start: rem(8px);
+}
+
+:host([active]) {
+    [part~='base'] {
+        background: var-get($theme, 'item-active-background');
+        color: var-get($theme, 'item-active-text-color');
+    }
+}
