+++ src/components/nav-drawer/themes/light/nav-drawer.material.scss
@@ -1,0 +1,51 @@
+@use 'styles/utilities' as *;
+@use 'themes' as *;
+
+$theme: $material;
+$background: var-get($theme, 'background') !default;
+
+:host {
+    @include css-vars-from-theme($theme, 'ig-nav-drawer');
+
+    [part='base'] {
+        border-radius: var-get($theme, 'border-radius');
+        background: $background;
+        border: none;
+        box-shadow: none;
+    }
+
+    [part='overlay'] {
+        background: color(gray, 500, .54);
+    }
+}
+
+:host([position='start']) [part='mini'],
+:host([position='end']) [part='mini'] {
+    background: $background;
+}
+
+:host([position='start']) [part='mini'] {
+    border-inline-end-color: color(gray, 900, .14);
+}
+
+:host([position='end']) [part='mini'] {
+    border-inline-start-color: color(gray, 900, .14);
+}
+
+:host([open][position='end']) {
+    [part='base'] {
+        border-inline-start-color: color(gray, 900, .14);
+    }
+}
+
+:host([position='relative']) {
+    [part='base'] {
+        border-inline-end-color: color(gray, 300);
+    }
+}
+
+:host([open]:not([position='relative'])) {
+    [part='base'] {
+        border-inline-end-color: color(gray, 900, .14);
+    }
+}
