+++ src/styles/themes/dark/bootstrap.scss
@@ -9,7 +9,6 @@
 @use '../../../components/tabs/themes/dark/tab.dark.bootstrap' as tab;
 @use '../../../components/tree/themes/dark/tree-item.bootstrap' as tree-item;
 @use '../../../components/select/themes/dark/select.bootstrap' as select;
-@use '../../../components/nav-drawer/themes/dark/nav-drawer-item.bootstrap' as nav-drawer-item;
 @use '../../../components/combo/themes/dark/combo.bootstrap' as combo;
 @use '../../../components/avatar/themes/dark/avatar.bootstrap' as avatar;
 @use '../../../components/button/themes/button/dark/button.bootstrap' as button;
@@ -19,6 +18,7 @@
 @use '../../../components/checkbox/themes/dark/switch.bootstrap' as switch;
 @use '../../../components/radio/themes/dark/radio.bootstrap' as radio;
 @use '../../../components/stepper/themes/step/dark/step.bootstrap' as stepper;
+@use '../../../components/nav-drawer/themes/dark/nav-drawer.bootstrap' as nav-drawer;
 @include base.root-styles($palette);
 @include dropdown.theme();
 @include input.theme();
@@ -26,7 +26,6 @@
 @include list.theme();
 @include tab.theme();
 @include tree-item.theme();
-@include nav-drawer-item.theme();
 @include select.theme();
 @include grid.theme();
 @include combo.theme();
@@ -38,3 +37,4 @@
 @include switch.theme();
 @include radio.theme();
 @include stepper.theme();
+@include nav-drawer.theme();
