+++ src/styles/themes/dark/indigo.scss
@@ -10,8 +10,6 @@
 @use '../../../components/list/themes/dark/list.indigo' as list;
 @use '../../../components/tree/themes/dark/tree-item.indigo' as tree-item;
 @use '../../../components/dropdown/themes/dark/dropdown.indigo' as dropdown;
-@use '../../../components/nav-drawer/themes/dark/nav-drawer.indigo' as nav-drawer;
-@use '../../../components/nav-drawer/themes/dark/nav-drawer-item.indigo' as nav-drawer-item;
 @use '../../../components/select/themes/dark/select.indigo' as select;
 @use '../../../components/snackbar/themes/dark/snackbar.indigo' as snackbar;
 @use '../../../components/combo/themes/dark/combo.indigo' as combo;
@@ -20,6 +18,7 @@
 @use '../../../components/navbar/themes/dark/navbar.indigo' as navbar;
 @use '../../../components/avatar/themes/dark/avatar.indigo' as avatar;
 @use '../../../components/stepper/themes/step/dark/step.indigo' as stepper;
+@use '../../../components/nav-drawer/themes/dark/nav-drawer.indigo' as nav-drawer;
 @include base.root-styles($palette);
 @include checkbox.theme();
 @include switch.theme();
@@ -29,8 +28,6 @@
 @include list.theme();
 @include tree-item.theme();
 @include dropdown.theme();
-@include nav-drawer.theme();
-@include nav-drawer-item.theme();
 @include select.theme();
 @include snackbar.theme();
 @include grid.theme();
@@ -40,3 +37,4 @@
 @include navbar.theme();
 @include avatar.theme();
 @include stepper.theme();
+@include nav-drawer.theme();
